@@ -1,4 +1,4 @@
-const { Router } = require('express')
+import { Router } from 'express'
 
 
 const routes = new Router()
@@ -7,5 +7,5 @@ routes.get('/' ,(req, res) => {
     return res.json( { ok: true })
 }) 
 
-module.exports = routes
+export default routes
 
